Cleanup warnings

+++ nlopt/timer.c
@@ -49,9 +49,9 @@ double nlopt_seconds(void)
 	  gettimeofday(&start, NULL);
      }
      gettimeofday(&tv, NULL);
-     return (tv.tv_sec - start.tv_sec) + 1.e-6 * (tv.tv_usec - start.tv_usec);
+     return (double)((tv.tv_sec - start.tv_sec) + 1.e-6 * (tv.tv_usec - start.tv_usec));
 #elif defined(HAVE_TIME)
-     return time(NULL);
+     return (double)time(NULL);
 #elif defined(_WIN32) || defined(__WIN32__)
      static THREADLOCAL ULONGLONG start;
      FILETIME ft;
@@ -61,7 +61,7 @@ double nlopt_seconds(void)
 	  start = (((ULONGLONG) ft.dwHighDateTime) << 32) + ft.dwLowDateTime;
      }
      GetSystemTimeAsFileTime(&ft);
-     return 100e-9 * (((((ULONGLONG) ft.dwHighDateTime) << 32) + ft.dwLowDateTime) - start);
+     return (double)(100e-9 * (((((ULONGLONG) ft.dwHighDateTime) << 32) + ft.dwLowDateTime) - start));
 #else
      /* use clock() as a fallback... this is somewhat annoying
 	because clock() may wrap around with a fairly short period */
@@ -70,7 +70,7 @@ double nlopt_seconds(void)
 	  start_inited = 1;
 	  start = clock();
      }
-     return (clock() - start) * 1.0 / CLOCKS_PER_SEC;
+     return (double)((clock() - start) * 1.0 / CLOCKS_PER_SEC);
 #endif
 }
 
@@ -80,14 +80,14 @@ unsigned long nlopt_time_seed(void)
 #if defined(HAVE_GETTIMEOFDAY)
      struct timeval tv;
      gettimeofday(&tv, NULL);
-     return (tv.tv_sec ^ tv.tv_usec);
+     return (unsigned long)(tv.tv_sec ^ tv.tv_usec);
 #elif defined(HAVE_TIME)
-     return time(NULL);
+     return (unsigned long)time(NULL);
 #elif defined(_WIN32) || defined(__WIN32__)
      FILETIME ft;
      GetSystemTimeAsFileTime(&ft);
-     return ft.dwHighDateTime ^ ft.dwLowDateTime;
+     return (unsigned long)ft.dwHighDateTime ^ ft.dwLowDateTime;
 #else
-     return clock();
+     return (unsigned long)clock();
 #endif
 }
